+++ superset/assets/src/visualizations/LineMulti/adaptor.jsx
@@ -1,4 +1,4 @@
-import createAdaptor from '../../../utils/createAdaptor';
+import createAdaptor from '../../utils/createAdaptor';
 import Component from './LineMulti';
 
 export default createAdaptor(Component);
